Move post box tests to component directory
